+++ fastapi/src/api/v1/__init__.py
@@ -1,4 +1,4 @@
 # app/api/endpoints/__init__.py
-from .films import router as films_router
-from .genres import router as genres_router
-from .persons import router as persons_router
+from .films_api import router as films_router
+from .genres_api import router as genres_router
+from .persons_api import router as persons_router
