+++ fastapi/src/services/genre_service.py
